fix sintax errors
