+++ src/actions/index.js
@@ -1,25 +1,15 @@
-const CREATE_BOOK = 'CREATE_BOOK';
-const REMOVE_BOOK = 'REMOVE_BOOK';
-
-const index = 0;
-
-const bookReducer = (state = {}, action) => {
-  switch(action.type){
-    case CREATE_BOOK:
-      return {
-        ...state,
-        newBook: {
-          id: ++index,
-          title: action.payload.title,
-          category: action.payload.category,
-        }
-      }
-      break;
-    case REMOVE_BOOK:
-      return state.map(book => { book.id != action.payload.id })
-      break;
-    default:
-      return state
-      break;
+const addBookAction = (book) => {
+  return {
+    type: 'CREATE_BOOK',
+    payload: book,
   }
 }
+
+const removeBookAction = (book) => {
+  return {
+    type: 'REMOVE_BOOK',
+    payload: book,
+  }
+}
+
+export { addBookAction, removeBookAction };
