+++ src/index.js
@@ -3,26 +3,27 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './components/App';
 import { Provider } from 'react-redux';
+import { createStore } from 'redux';
 import booksReducer from './reducers/books';
 
 const defaultState = {
-    {
-      id: Math.random(Math.random() * 100),
-      title: 'Learn programming the hard way',
-      category: 'Programming',
-    },
-    {
-      id: Math.random(Math.random() * 100),
-      title: 'The Lean Startup',
-      category: 'Novel',
-    },
+  firstBook: {
+    id: Math.random(Math.random() * 100),
+    title: 'Learn programming the hard way',
+    category: 'Programming',
+  },
+  secondBook: {
+    id: Math.random(Math.random() * 100),
+    title: 'The Lean Startup',
+    category: 'Novel',
+  },
 };
 
 const store = createStore(booksReducer, defaultState);
 
 ReactDOM.render(
   <React.StrictMode>
-    <Provider>
+    <Provider store={store}>
       <App />
     </Provider>
   </React.StrictMode>,
